+++ src/lib.rs
@@ -66,6 +66,7 @@ pub mod audio_mix;
 pub mod duration;
 pub mod id;
 pub mod light;
+pub mod material;
 pub mod object;
 pub mod ops;
 pub mod page;
@@ -90,6 +91,10 @@ pub use audio_mix::mix_cues;
 pub use duration::{Lifetime, SceneDuration, TimeStamp};
 pub use id::ObjectId;
 pub use light::{Light, LightCommon, LightInstance, SpotParams};
+pub use material::{
+    AlphaMode, Material, NormalTextureBinding, OcclusionTextureBinding, PbrMetallicRoughness,
+    TextureBinding,
+};
 pub use object::{
     BlendMode, Canvas, ClipRect, Effect, ImageSource, LengthUnit, LiveStreamHandle, ObjectKind,
     Sample, SceneObject, Shape, TextRun, Transform, VideoSource,